Initial Commit
A program in C that inputs and stores student details of a school and retrieves the information when required

This program uses a struct to store the details of each student, including the name, age, and roll number. The input() function prompts the user to enter the number of students and then collects the details for each student. The display() function takes a roll number as an argument and searches for the student with that roll number in the struct, printing their details if found.